Ship 1.0.10 through 1.1.0, and add Minecraft 26.3
Minecraft 26.3 lands on Fabric and NeoForge, taking the matrix to 60 nodes
(23 fabric / 18 forge / 19 neoforge). Forge published no 26.3 build, so 26.3
is a two-loader band, the same shape as 1.21.2.

The 26.3 work, in the order it was found:

- Biome resolution split into two factories. 26.3 added BiomeSource#createResolverForChunk
  beside #createResolver, and MultiNoiseBiomeSource overrides it without routing
  through the first. Wrapping only the original meant /locate and cave maps saw the
  mod's placement while generation did not, so AC biomes were located but never
  generated. A second @Inject now shares one resolver wrapper.
- Five client crashes a green build could not see: the colour EnvironmentAttributes
  retyped to joml vectors, BindGroupLayouts.MATRICES_PROJECTION splitting rather than
  being renamed, the deleted RenderSystem output-texture overrides, RenderType#prepare()
  having to run before the pass is opened, and InterpolationHandler becoming an
  interface with a nullable PositionPath.
- NeoForge 26.3.0.4-beta deleted five APIs this tree calls, so 26.3 is the only node
  that declares a NeoForge floor.

Also folds in the 1.0.10 and 1.0.11 releases: the worldgen lake crash, the Abyssal
Chasm colour overflow, Fabric fluid render handlers, the compendium and spelunkery
text going invisible from 1.21.6, the Forge 26.2 access-transformer parser, and the
rest of the player-reported fixes from those waves.

Working material (development notes, player reports, screenshots, audit and release
tooling, and the generated per-node trees) is no longer tracked. The build system
itself stays in the repository.

+++ src/main/java/com/github/alexmodguy/alexscaves/server/level/biome/BiomeSampler.java
@@ -59,7 +59,9 @@ public class BiomeSampler {
         double d1 = quartPos.y() - (double)j;
         double d2 = quartPos.z() - (double)k;
         double d3 = 0.0D;
-        float f = 0;
+        // Summed as a double, like the weights: a float running total drifts past the weight sum, so a
+        // column that is 100% one biome could come back as 1.0000001 and a 0 as a hair below zero.
+        double f = 0;
         BiomeManager biomeManager = level.getBiomeManager();
 
         for(int l = 0; l < 6; ++l) {
@@ -80,7 +82,7 @@ public class BiomeSampler {
             }
         }
 
-        return f / (float) d3;
+        return (float) (f / d3);
     }
 
     @FunctionalInterface
